add service.py

+++ default.py
@@ -2,6 +2,7 @@
 
 from __future__ import unicode_literals
 
+import xbmc
 import xbmcgui
 import xbmcplugin
 import xbmcaddon
@@ -15,20 +16,20 @@ import email.utils
 import time, datetime
 
 from bs4 import BeautifulSoup
-from resources.lib.common import log, notify
 
+from resources.lib.common import log, notify
 from resources.lib.gmail import Gmail
 from resources.lib.icloud import iCloud
 
 class Main:
 
-    def __init__(self):
+    def __init__(self, service=None):
 
         # アドオン
         self.addon = xbmcaddon.Addon()
 
         # メールサービスを初期化
-        service = self.addon.getSetting('service')
+        service = service or self.addon.getSetting('service')
         if service == 'Gmail':
             address = self.addon.getSetting('address1')
             password = self.addon.getSetting('password1')
@@ -49,15 +50,15 @@ class Main:
             notify('unknown service: %s' % service, error=True)
             sys.exit()
 
-
         # ファイル/ディレクトリパス
         self.profile_path = xbmc.translatePath(self.addon.getAddonInfo('profile').decode('utf-8'))
         self.cache_path = os.path.join(self.profile_path, 'cache', service)
         if not os.path.isdir(self.cache_path):
             os.makedirs(self.cache_path)
 
+    def main(self):
         # パラメータ抽出
-        params = {'action':'','filename':'','subject':'','message':'', 'to':'', 'cc':'', 'bcc':''}
+        params = {'action':'','filename':'','subject':'','message':'', 'to':'', 'cc':''}
         args = urlparse.parse_qs(sys.argv[2][1:])
         for key in params.keys():
             params[key] = args.get(key, None)
@@ -66,28 +67,8 @@ class Main:
 
         # メイン処理
         if params['action'] is None:
-            # メール取得
-            # 前回取得の月日を読み込む
-            criterion_file = os.path.join(self.cache_path, '.criterion')
-            if os.path.isfile(criterion_file):
-                f = open(criterion_file, 'r')
-                criterion = f.read().decode('utf-8')
-                f.close()
-            else:
-                d = datetime.datetime.now() - datetime.timedelta(days=30)
-                criterion = d.strftime('SINCE %d-%b-%Y').decode('utf-8')
-            lastid, n = self.receive(criterion)
-            self.list()
-            # 取得されたメール数を通知
-            notify('%d mails retrieved' % n)
-            # 次回取得のために月日を保存
-            #d = datetime.datetime.now()
-            #criterion = d.strftime('SINCE %d-%b-%Y')
-            if lastid:
-                criterion = '%d:%d' % (int(lastid)+1,int(lastid)+200)
-                f = open(criterion_file, 'w')
-                f.write(criterion)
-                f.close()
+            # メールをチェック
+            self.check()
         elif params['action'] == 'refresh':
             # キャッシュクリア
             files = os.listdir(self.cache_path)
@@ -99,10 +80,63 @@ class Main:
             # メールの内容を表示
             filename = urllib.unquote_plus(params['filename'])
             if filename: self.open(filename)
+        elif params['action'] == 'send':
+            # メールを送信
+            if self.addon.getSetting('bcc') == 'true':
+                bcc = [self.service.smtp_from]
+            else:
+                bcc = None
+            self.send(subject=params['subject'], message=params['message'], to=params['to'], cc=params['cc'], bcc=bcc)
 
-    def send(self, subject, message, to, cc=None, replyto=None):
+    def check(self, silent=False):
+        # 管理用ファイル
+        criterion_file = os.path.join(self.cache_path, '.criterion')
+        newmails_file = os.path.join(self.cache_path, '.newmails')
+        # 前回表示時の月日を読み込む
+        if os.path.isfile(criterion_file):
+            f = open(criterion_file, 'r')
+            criterion = f.read().decode('utf-8')
+            f.close()
+        else:
+            # 前回表示時の月日が不明の場合は30日前に設定
+            d = datetime.datetime.utcnow() - datetime.timedelta(days=30)
+            criterion = d.strftime('SINCE %d-%b-%Y').decode('utf-8')
+        # 設定した期間のメールを検索
+        newmails = self.receive(criterion)
+        if len(newmails) > 0:
+            # 新着メールのファイルパスリストを書き出す
+            f = open(newmails_file, 'a')
+            f.write('\n'.join(newmails) + '\n')
+            f.close()
+            xbmc.sleep(1000)
+            # Kodiをアクティベート
+            if self.addon.getSetting('cec') == 'true':
+                xbmc.executebuiltin('XBMC.CECActivateSource')
+            # 新着があることを通知
+            notify('New mail arrived')
+        # アドオン操作で呼び出された場合の処理
+        if silent == False:
+            if os.path.isfile(newmails_file):
+                # 新着メールのファイルパスリストを読み込む
+                f = open(newmails_file, 'r')
+                newmails = f.read().split('\n')
+                f.close()
+                # 新着メールのファイルパスリストを削除
+                os.remove(newmails_file)
+            else:
+                newmails = []
+            # リストを表示
+            self.list(newmails)
+            # 次回表示時のために今回表示時の月日を書き出す
+            d = datetime.datetime.utcnow()
+            criterion = d.strftime('SINCE %d-%b-%Y')
+            f = open(criterion_file, 'w')
+            f.write(criterion)
+            f.close()
+
+    def send(self, subject, message, to, cc=None, bcc=None):
         # メール送信
-        self.service.send(subject, message, [to], cc, replyto_address=replyto)
+        self.service.send(subject, message, to, cc, bcc)
         # 通知
         notify('Message has been sent to %s' % to)
 
@@ -119,7 +153,7 @@ class Main:
         if re.compile(r'<html|<!doctype', re.IGNORECASE).match(decoded_body):
             decoded_body = re.compile(r'<style.*?</style>', re.DOTALL|re.MULTILINE|re.IGNORECASE).sub('',decoded_body)
             decoded_body = re.compile(r'<script.*?</script>', re.DOTALL|re.MULTILINE|re.IGNORECASE).sub('',decoded_body)
-            soup = BeautifulSoup(decoded_body)
+            soup = BeautifulSoup(decoded_body, 'html.parser')
             buf = []
             for text in soup.stripped_strings:
                 buf.append(text)
@@ -128,31 +162,36 @@ class Main:
 
     def receive(self, criterion):
         # メール取得
-        lastid, mails = self.service.receive(criterion, 'TEXT')
+        mails = self.service.receive(criterion, 'TEXT')
+        # 新規メールのリスト
+        newmails = []
         # メール毎の処理
         for mail in mails:
             # データ変換
             filename, timestamp = self.convert(mail)
-            # ファイル書き込み
+            # ファイルパス
             filepath = os.path.join(self.cache_path, filename)
-            f = codecs.open(filepath,'w','utf-8')
-            # header
-            for key in ['From','To','CC','Date','Subject','Message-Id']:
-                if mail[key]:
-                    f.write('%s: ' % key)
-                    f.write(mail[key].replace('\r\n',''))
-                    f.write('\r\n')
-            # separator
-            f.write('\r\n')
-            # body
-            f.write(mail['body'])
-            # ファイル書き込み完了
-            f.close()
-            # タイムスタンプを変更
-            os.utime(filepath, (timestamp, timestamp))
-        return lastid, len(mails)
+            if not os.path.isfile(filepath):
+                content = ''
+                # header
+                for key in ['From','To','CC','Date','Subject','Message-Id']:
+                    if mail[key]: content += '%s: %s\r\n' % (key, mail[key].replace('\r\n',''))
+                # separator
+                content += '\r\n'
+                # body
+                content += mail['body']
+                # ファイル書き込み
+                f = codecs.open(filepath,'w','utf-8')
+                f.write(content)
+                f.close()
+                # タイムスタンプを変更
+                os.utime(filepath, (timestamp, timestamp))
+                # 新規メールのパスを格納
+                newmails.append(filepath)
+        log('%d/%d mails retrieved using criterion "%s"' % (len(newmails),len(mails),criterion))
+        return newmails
 
-    def list(self):
+    def list(self, newmails):
         #ファイルリスト
         files = sorted(os.listdir(self.cache_path), key=lambda mail: os.stat(os.path.join(self.cache_path,mail)).st_mtime, reverse=True)
         #ファイルの情報を表示
@@ -184,6 +223,8 @@ class Main:
                         title = params['Subject']
                     else:
                         title = self.addon.getLocalizedString(30901)
+                    if filepath in newmails:
+                        title = '[COLOR yellow]%s[/COLOR]' % title
                     plot = '[COLOR green]Date:[/COLOR] %s\n[COLOR green]From:[/COLOR] %s' % (params['Date'],params['From'])
                     query = '%s?action=open&filename=%s' % (sys.argv[0],urllib.quote_plus(filename))
                     listitem = xbmcgui.ListItem(title, iconImage="DefaultFolder.png", thumbnailImage="DefaultFolder.png")
@@ -243,4 +284,4 @@ class Main:
         viewer.getControl(5).setText('%s\n\n%s' % ('\n'.join(header),'\n'.join(body)))
 
 
-if __name__  == '__main__': Main()
+if __name__  == '__main__': Main().main()
